+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Gradle to version 8.13 [#343](https://github.com/IN-CORE/incore-services/issues/343)
 - Jersey to version 3.1.10 [#340](https://github.com/IN-CORE/incore-services/issues/340)
 - Swagger to version 2.2.28 [#251](https://github.com/IN-CORE/incore-services/issues/251)
+- Endpoint to delete and edit layers in visualization [#360](https://github.com/IN-CORE/incore-services/issues/360)
 
 ## [1.28.0] - 2025-02-20
 ### Added